refactor: make jsonschema package public
