+++ internal/oas/parameter.go
@@ -1,6 +1,6 @@
 package oas
 
-import "github.com/ogen-go/ogen/internal/jsonschema"
+import "github.com/ogen-go/ogen/jsonschema"
 
 // ParameterLocation defines where OpenAPI parameter is located.
 type ParameterLocation string
