+++ internal/oas/parser/parse_request_body.go
@@ -4,8 +4,8 @@ import (
 	"github.com/go-faster/errors"
 
 	"github.com/ogen-go/ogen"
-	"github.com/ogen-go/ogen/internal/jsonschema"
 	"github.com/ogen-go/ogen/internal/oas"
+	"github.com/ogen-go/ogen/jsonschema"
 )
 
 func (p *parser) parseRequestBody(body *ogen.RequestBody) (*oas.RequestBody, error) {
