+++ jsonschema/parser_enum.go
