+++ jsonschema/schema.go
